footer 폴더 위치변경 리뷰 반영
