--- src/pages/Main/Components/Footer/Footer.scss
@@ -1,130 +1,0 @@
-.footerContainer {
-  width: 100%;
-  bottom: 0;
-  .footerLoginContainer {
-    height: 63px;
-    width: 100%;
-    background-color: #67625d;
-    border: 1px solid black;
-    .footerLogin {
-      height: 100%;
-      margin: 0 0 0 75%;
-      display: flex;
-      align-items: center;
-      .footerLoginBtn {
-        height: 40px;
-        width: 94px;
-        background-color: #3f3e3c;
-        color: #cccccc;
-        border: 1px solid black;
-      }
-      .goToSns {
-        display: flex;
-        justify-content: space-around;
-        width: 250px;
-        .facebook {
-          font-size: 35px;
-          padding: 5px;
-          background-color: #3f3e3c;
-          border-radius: 100%;
-          color: #cccccc;
-        }
-        .insta {
-          font-size: 35px;
-          padding: 5px;
-          background-color: #3f3e3c;
-          border-radius: 100%;
-          color: #cccccc;
-        }
-        .youtube {
-          font-size: 35px;
-          padding: 5px;
-          background-color: #3f3e3c;
-          color: #cccccc;
-          border-radius: 100%;
-        }
-      }
-    }
-  }
-
-  .companyInfo {
-    display: flex;
-    justify-content: space-around;
-    align-items: center;
-    height: 140px;
-    width: 100%;
-    background-color: #fafafa;
-    .iconConnect {
-      display: flex;
-      width: 400px;
-      justify-content: space-around;
-      border-left: 1px solid #595959;
-      .column {
-        display: flex;
-        flex-direction: column;
-        align-items: center;
-        .abcd {
-          font-size: 45px;
-          margin-bottom: 5px;
-        }
-      }
-    }
-    .contentWrapper {
-      display: flex;
-      justify-content: space-around;
-
-      .callCenterInfoBox {
-        display: flex;
-        flex-direction: column;
-        .gray {
-          color: #7b7b7b;
-        }
-        .companyNum {
-          font-size: #595959;
-          font-size: 33px;
-        }
-      }
-      .buyingRequestBox {
-        display: flex;
-        flex-direction: column;
-        margin-left: 30px;
-        .email {
-          color: #595959;
-          font-size: 18px;
-        }
-        .kakaoId {
-          color: #797979;
-          font-size: 13px;
-        }
-        .gray {
-          color: #a0a0a0;
-        }
-      }
-    }
-  }
-  .companyDetail {
-    width: 684px;
-    height: 60px;
-    font-size: 12px;
-    display: flex;
-    margin: 40px;
-    .footerLogo {
-      font-size: 30px;
-      margin: auto;
-    }
-    .gray {
-      color: #a0a0a0;
-      .companyAddress {
-        margin-bottom: 10px;
-      }
-    }
-  }
-  .footerLeft {
-    .fontsize14 {
-      font-size: 14px;
-    }
-  }
-  .footerLeft {
-    margin-left: 184px;
-  }
-}
